Move SourceType decorators into the source package
