+++ connectors/activemq/src/test/java/forklift/activemq/test/MessagingTest.java
@@ -11,10 +11,10 @@ import forklift.decorators.LifeCycle;
 import forklift.decorators.OnMessage;
 import forklift.decorators.OnValidate;
 import forklift.decorators.Producer;
-import forklift.decorators.Queue;
 import forklift.message.Header;
 import forklift.producers.ForkliftProducerI;
 import forklift.producers.ProducerException;
+import forklift.source.decorators.Queue;
 
 import org.apache.activemq.command.ActiveMQTextMessage;
 import org.junit.After;
