+++ connectors/activemq/src/test/java/forklift/activemq/test/ProducerTest.java
@@ -11,8 +11,8 @@ import forklift.decorators.LifeCycle;
 import forklift.decorators.OnMessage;
 import forklift.decorators.OnValidate;
 import forklift.decorators.Producer;
-import forklift.decorators.Queue;
-import forklift.decorators.Topic;
+import forklift.source.decorators.Queue;
+import forklift.source.decorators.Topic;
 import forklift.message.Header;
 import forklift.producers.ForkliftProducerI;
 import forklift.producers.ProducerException;
