+++ connectors/kafka/src/test/java/forklift/integration/ForkliftMessageTests.java
@@ -1,20 +1,23 @@
 package forklift.integration;
 
 import static org.junit.Assert.assertTrue;
+
 import forklift.Forklift;
 import forklift.connectors.ConnectorException;
 import forklift.connectors.ForkliftMessage;
 import forklift.consumer.Consumer;
 import forklift.decorators.OnMessage;
 import forklift.decorators.Producer;
-import forklift.decorators.Queue;
 import forklift.exception.StartupException;
 import forklift.integration.server.TestServiceManager;
 import forklift.producers.ForkliftProducerI;
 import forklift.producers.ProducerException;
+import forklift.source.decorators.Queue;
+
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
+
 import java.util.HashMap;
 import java.util.Map;
