+++ consumers/forklift-jarjar-test/src/main/java/forklift/consumer/JarJarConsumer.java
@@ -1,7 +1,7 @@
 package forklift.consumer;
 
 import forklift.decorators.OnMessage;
-import forklift.decorators.Queue;
+import forklift.source.decorators.Queue;
 
 @Queue("test")
 public class JarJarConsumer {
