+++ consumers/forklift-multitq-test/src/main/java/forklift/consumer/MultiQConsumer.java
@@ -1,7 +1,7 @@
 package forklift.consumer;
 
 import forklift.decorators.OnMessage;
-import forklift.decorators.Queue;
+import forklift.source.decorators.Queue;
 
 @Queue("test1")
 public class MultiQConsumer {
