+++ consumers/forklift-test/src/main/java/forklift/consumer/TestQueueConsumer.java
@@ -1,7 +1,7 @@
 package forklift.consumer;
 
 import forklift.decorators.OnMessage;
-import forklift.decorators.Queue;
+import forklift.source.decorators.Queue;
 
 @Queue("test")
 public class TestQueueConsumer {
