+++ core/src/main/java/forklift/source/decorators/Topic.java
@@ -1,4 +1,4 @@
-package forklift.decorators;
+package forklift.source.decorators;
 
 import forklift.source.TopicSource;
