+++ core/src/test/java/forklift/consumer/OnDecoratorTest.java
@@ -10,7 +10,8 @@ import forklift.connectors.ForkliftMessage;
 import forklift.decorators.On;
 import forklift.decorators.OnMessage;
 import forklift.decorators.OnValidate;
-import forklift.decorators.Queue;
+import forklift.source.decorators.Queue;
+
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
