+++ core/src/test/java/forklift/deployment/deploymentClasses/TestQueue1.java
@@ -1,6 +1,6 @@
 package forklift.deployment.deploymentClasses;
 
-import forklift.decorators.Queue;
+import forklift.source.decorators.Queue;
 
 /**
  * Created by afrieze on 10/31/16.
