+++ core/src/test/java/forklift/deployment/deploymentClasses/TestTopic1.java
@@ -1,6 +1,6 @@
 package forklift.deployment.deploymentClasses;
 
-import forklift.decorators.Topic;
+import forklift.source.decorators.Topic;
 
 /**
  * Created by afrieze on 10/31/16.
