+++ plugins/replay/src/main/java/forklift/replay/ReplayConsumer.java
@@ -5,7 +5,7 @@ import forklift.decorators.MultiThreaded;
 import forklift.decorators.OnMessage;
 import forklift.decorators.OnValidate;
 import forklift.decorators.Order;
-import forklift.decorators.Queue;
+import forklift.source.decorators.Queue;
 
 import javax.inject.Inject;
